merl_transform (syntax_tools v3.2)

Parse transform for merl. Enables the use of automatic metavariables and using quasi-quotes in matches and case switches. Also optimizes calls to functions in merl by partially evaluating them, turning strings to templates, etc., at compile-time.

Using -include_lib("syntax_tools/include/merl.hrl"). enables this transform, unless the macro MERL_NO_TRANSFORM is defined first.

Functions

parse_transform(InForms, Options)

-spec parse_transform(InForms, Options :: term()) -> OutForms
                   when
                       InForms :: [erl_parse:abstract_form() | erl_parse:form_info()],
                       OutForms :: [erl_parse:abstract_form() | erl_parse:form_info()].